Utilização do ApiContext para centralizar persistência
Padronização da forma de utilização. Controle transacional não é mais feito no filtro
