+++ siga/src/main/java/br/gov/jfrj/siga/api/v1/ApiContext.java
@@ -1,9 +1,13 @@
 package br.gov.jfrj.siga.api.v1;
 
+import static java.util.Objects.isNull;
+
 import java.io.Closeable;
 import java.io.IOException;
-
 import javax.persistence.EntityManager;
+
+import com.crivano.swaggerservlet.SwaggerAuthorizationException;
+import com.crivano.swaggerservlet.SwaggerServlet;
 
 import br.gov.jfrj.siga.base.log.RequestLoggerFilter;
 import br.gov.jfrj.siga.cp.bl.Cp;
@@ -11,13 +15,18 @@ import br.gov.jfrj.siga.dp.dao.CpDao;
 import br.gov.jfrj.siga.model.ContextoPersistencia;
 import br.gov.jfrj.siga.model.dao.ModeloDao;
 import br.gov.jfrj.siga.util.SigaStarter;
+import br.gov.jfrj.siga.vraptor.SigaObjects;
 
 public class ApiContext implements Closeable {
 	EntityManager em;
 	boolean transacional;
 	long inicio = System.currentTimeMillis();
 
-	public ApiContext(boolean transacional) {
+	public ApiContext(boolean transacional, boolean autenticar) throws SwaggerAuthorizationException {
+		if (autenticar) {
+			buscarEValidarUsuarioLogado();
+		}
+		
 		this.transacional = transacional;
 		em = SigaStarter.emf.createEntityManager();
 		ContextoPersistencia.setEntityManager(em);
@@ -55,4 +64,44 @@ public class ApiContext implements Closeable {
 			ContextoPersistencia.setEntityManager(null);
 		}
 	}
+
+	/**
+	 * Retorna uma instância de {@link SigaObjects} a partir do Request do
+	 * {@link SwaggerServlet}.
+	 * @throws Exception Se houver algo de errado.
+	 */
+	static SigaObjects getSigaObjects() throws Exception {
+		SigaObjects sigaObjects = new SigaObjects(SwaggerServlet.getHttpServletRequest());
+		return sigaObjects;
+	}
+
+	/**
+	 * Verifica a presença de um usuário logado e o retorna.
+	 * 
+	 * @return O login do Usuário na sessão
+	 * @throws SwaggerAuthorizationException Se não achar nenhum usuário logado na
+	 *                                       sessão.
+	 * @see ContextoPersistencia#getUserPrincipal()
+	 */
+	static String buscarEValidarUsuarioLogado() throws SwaggerAuthorizationException {
+		String userPrincipal = ContextoPersistencia.getUserPrincipal();
+		if (isNull(userPrincipal)) {
+			throw new SwaggerAuthorizationException("Usuário não está logado");
+		}
+
+		return userPrincipal;
+	}
+
+	/**
+	 * Verifica se o usuário tem acesso ao sistema e ao serviço informado 
+	 * no parâmetro acesso.
+	 * 
+	 * @param acesso              Caminho do serviço a ser verificado a permissão de acesso
+	 * 
+	 * @throws Exception Se houver algo de errado.
+	 */
+	static void assertAcesso(String acesso) throws Exception {
+		ApiContext.getSigaObjects().assertAcesso(acesso);
+	}
+
 }
