+++ siga/src/main/java/br/gov/jfrj/siga/api/v1/ISigaApiV1.java
@@ -1,14 +1,11 @@
 package br.gov.jfrj.siga.api.v1;
 
-import java.io.InputStream;
+import java.util.Date;
 import java.util.List;
-import java.util.Map;
-
 import com.crivano.swaggerservlet.ISwaggerMethod;
 import com.crivano.swaggerservlet.ISwaggerModel;
 import com.crivano.swaggerservlet.ISwaggerRequest;
 import com.crivano.swaggerservlet.ISwaggerResponse;
-import com.crivano.swaggerservlet.ISwaggerResponseFile;
 
 public interface ISigaApiV1 {
 	public class Status implements ISwaggerModel {
@@ -74,6 +71,11 @@ public interface ISigaApiV1 {
 		public String idpai;
 	}
 
+	public class AcessoItem implements ISwaggerModel {
+		public Date datahora;
+		public String ip;
+	}
+
 	public class AutenticarPostRequest implements ISwaggerRequest {
 	}
 
@@ -83,89 +85,6 @@ public interface ISigaApiV1 {
 
 	public interface IAutenticarPost extends ISwaggerMethod {
 		public void run(AutenticarPostRequest req, AutenticarPostResponse resp) throws Exception;
-	}
-
-	public class DocumentosSiglaArquivoGetRequest implements ISwaggerRequest {
-		public String contenttype;
-		public String sigla;
-		public Boolean estampa;
-		public Boolean completo;
-		public Boolean volumes;
-	}
-
-	public class DocumentosSiglaArquivoGetResponse implements ISwaggerResponse {
-		public String uuid;
-		public String jwt;
-	}
-
-	public interface IDocumentosSiglaArquivoGet extends ISwaggerMethod {
-		public void run(DocumentosSiglaArquivoGetRequest req, DocumentosSiglaArquivoGetResponse resp) throws Exception;
-	}
-
-	public class DownloadJwtFilenameGetRequest implements ISwaggerRequest {
-		public String jwt;
-		public String filename;
-		public String disposition;
-	}
-
-	public class DownloadJwtFilenameGetResponse implements ISwaggerResponse, ISwaggerResponseFile {
-		public String contenttype = "application/pdf";
-		public String contentdisposition = "attachment";
-		public Long contentlength;
-		public InputStream inputstream;
-		public Map<String, List<String>> headerFields;
-
-		public String getContenttype() {
-			return contenttype;
-		}
-		public void setContenttype(String contenttype) {
-			this.contenttype = contenttype;
-		}
-		public String getContentdisposition() {
-			return contentdisposition;
-		}
-		public void setContentdisposition(String contentdisposition) {
-			this.contentdisposition = contentdisposition;
-		}
-		public Long getContentlength() {
-			return contentlength;
-		}
-		public void setContentlength(Long contentlength) {
-			this.contentlength = contentlength;
-		}
-		public InputStream getInputstream() {
-			return inputstream;
-		}
-		public void setInputstream(InputStream inputstream) {
-			this.inputstream = inputstream;
-		}
-		public Map<String, List<String>> getHeaderFields() {
-			return headerFields;
-		}
-		public void setHeaderFields(Map<String, List<String>> headerFields) {
-			this.headerFields = headerFields;
-		}
-	}
-
-	public interface IDownloadJwtFilenameGet extends ISwaggerMethod {
-		public void run(DownloadJwtFilenameGetRequest req, DownloadJwtFilenameGetResponse resp) throws Exception;
-	}
-
-	public class StatusChaveGetRequest implements ISwaggerRequest {
-		public String chave;
-	}
-
-	public class StatusChaveGetResponse implements ISwaggerResponse {
-		public String mensagem;
-		public Double indice;
-		public Double contador;
-		public Double bytes;
-		public String errormsg;
-		public String stacktrace;
-	}
-
-	public interface IStatusChaveGet extends ISwaggerMethod {
-		public void run(StatusChaveGetRequest req, StatusChaveGetResponse resp) throws Exception;
 	}
 
 	public class PessoasGetRequest implements ISwaggerRequest {
@@ -193,6 +112,30 @@ public interface ISigaApiV1 {
 
 	public interface ILotacoesGet extends ISwaggerMethod {
 		public void run(LotacoesGetRequest req, LotacoesGetResponse resp) throws Exception;
+	}
+
+	public class TokenCriarPostRequest implements ISwaggerRequest {
+		public String username;
+		public String password;
+	}
+
+	public class TokenCriarPostResponse implements ISwaggerResponse {
+		public String id_token;
+	}
+
+	public interface ITokenCriarPost extends ISwaggerMethod {
+		public void run(TokenCriarPostRequest req, TokenCriarPostResponse resp) throws Exception;
+	}
+
+	public class AcessosGetRequest implements ISwaggerRequest {
+	}
+
+	public class AcessosGetResponse implements ISwaggerResponse {
+		public List<AcessoItem> list;
+	}
+
+	public interface IAcessosGet extends ISwaggerMethod {
+		public void run(AcessosGetRequest req, AcessosGetResponse resp) throws Exception;
 	}
 
 }
