+++ sigaex/src/main/java/br/gov/jfrj/siga/ex/api/v1/DocumentosSiglaArquivoGet.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpServletRequest;
 
 import com.crivano.swaggerservlet.PresentableUnloggedException;
 import com.crivano.swaggerservlet.SwaggerAuthorizationException;
+import com.crivano.swaggerservlet.SwaggerException;
 import com.crivano.swaggerservlet.SwaggerServlet;
 
 import br.gov.jfrj.itextpdf.Status;
@@ -15,7 +16,9 @@ import br.gov.jfrj.siga.ex.ExMobil;
 import br.gov.jfrj.siga.ex.api.v1.IExApiV1.DocumentosSiglaArquivoGetRequest;
 import br.gov.jfrj.siga.ex.api.v1.IExApiV1.DocumentosSiglaArquivoGetResponse;
 import br.gov.jfrj.siga.ex.api.v1.IExApiV1.IDocumentosSiglaArquivoGet;
+import br.gov.jfrj.siga.ex.bl.CurrentRequest;
 import br.gov.jfrj.siga.ex.bl.Ex;
+import br.gov.jfrj.siga.ex.bl.RequestInfo;
 import br.gov.jfrj.siga.hibernate.ExDao;
 import br.gov.jfrj.siga.model.ContextoPersistencia;
 import br.gov.jfrj.siga.persistencia.ExMobilDaoFiltro;
@@ -26,7 +29,9 @@ public class DocumentosSiglaArquivoGet implements IDocumentosSiglaArquivoGet {
 
 	@Override
 	public void run(DocumentosSiglaArquivoGetRequest req, DocumentosSiglaArquivoGetResponse resp) throws Exception {
-		try (ApiContext ctx = new ApiContext(false)) {
+		try (ApiContext ctx = new ApiContext(false, false)) {
+			CurrentRequest.set(
+					new RequestInfo(null, SwaggerServlet.getHttpServletRequest(), SwaggerServlet.getHttpServletResponse()));
 			String usuario = ContextoPersistencia.getUserPrincipal();
 
 			if (usuario == null)
@@ -53,6 +58,11 @@ public class DocumentosSiglaArquivoGet implements IDocumentosSiglaArquivoGet {
 			final String contextpath = request.getContextPath();
 
 			iniciarGeracaoDePdf(req, resp, usuario, filename, contextpath, servernameport);
+		} catch (AplicacaoException | SwaggerException e) {
+			throw e;
+		} catch (Exception e) {
+			e.printStackTrace(System.out);
+			throw e;
 		}
 	}
