+++ sigaex/src/main/java/br/gov/jfrj/siga/ex/api/v1/DocumentosSiglaHtmlGet.java
@@ -1,8 +1,11 @@
 package br.gov.jfrj.siga.ex.api.v1;
 
+import java.util.Base64;
+import java.util.Base64.Decoder;
 import java.util.Map;
 
 import org.apache.http.HttpHeaders;
+import org.codehaus.jettison.json.JSONObject;
 
 import com.auth0.jwt.JWTVerifier;
 import com.crivano.swaggerservlet.SwaggerException;
@@ -22,66 +25,50 @@ import br.gov.jfrj.siga.ex.util.ProcessadorHtml;
 import br.gov.jfrj.siga.hibernate.ExDao;
 import br.gov.jfrj.siga.persistencia.ExMobilDaoFiltro;
 
-@AcessoPublico
+@AcessoPublicoEPrivado
 public class DocumentosSiglaHtmlGet implements IDocumentosSiglaHtmlGet {
 	@Override
 	public void run(DocumentosSiglaHtmlGetRequest req, DocumentosSiglaHtmlGetResponse resp) throws Exception {
-		try {
-			CurrentRequest.set(new RequestInfo(null, SwaggerServlet.getHttpServletRequest(), SwaggerServlet.getHttpServletResponse()));
+		try (ApiContext ctx = new ApiContext(false, false)) {
+			CurrentRequest.set(
+					new RequestInfo(null, SwaggerServlet.getHttpServletRequest(), SwaggerServlet.getHttpServletResponse()));
 	
 	        String jwt = CurrentRequest.get().getRequest().getHeader(HttpHeaders.AUTHORIZATION);		
-			if (jwt != null) {
+			ExMobilDaoFiltro flt = new ExMobilDaoFiltro();
+			flt.setSigla(req.sigla);
+			ExMobil mob = ExDao.getInstance().consultarPorSigla(flt);
+			if (mob == null) {
+				throw new SwaggerException("Documento não encontrado: " + req.sigla, 404, null, req, resp,
+						null);
+			}
+			Decoder decoder = Base64.getUrlDecoder();
+		    String[] jwt_split = jwt.split("\\.");
+		    String jwtBody = new String(decoder.decode(jwt_split[1]));
+		    JSONObject jwtBodyJson = new JSONObject(jwtBody);
+
+		    if (jwtBodyJson.has("n")) {
 				String n = verifyJwtToken(jwt).get("n").toString();
 				ExProtocolo protocolo = ExDao.getInstance().obterProtocoloPorCodigo(n);
 				ExDocumento docPai = protocolo.getExDocumento();
-						
-				ExMobilDaoFiltro flt = new ExMobilDaoFiltro();
-				flt.setSigla(req.sigla);
-				ExMobil mob = ExDao.getInstance().consultarPorSigla(flt);
-				if (mob == null) {
-					throw new SwaggerException("Documento não encontrado: " + req.sigla, 404, null, req, resp,
-							null);
-				}
-				ExDocumento doc = mob.doc();
-	
 				if (!(docPai.getIdDoc() == mob.getExMobilPai().getDoc().getIdDoc()
 						&& mob.getPodeExibirNoAcompanhamento())) {
 					throw new SwaggerException("Documento não permitido para visualização: " + req.sigla, 403, null, req, resp,
 							null);
 				}
-				resp.html = ProcessadorHtml.bodyOnly(doc.getHtml());
-				return;
 			} else {
-				throw new SwaggerException("O token de acesso não foi enviado." + req.sigla, 401, null, req, resp,
-						null);
+				ApiContext.buscarEValidarUsuarioLogado();
+				ApiContext.assertAcesso("");
 			}
+			ExDocumento doc = mob.doc();
+
+			resp.html = ProcessadorHtml.bodyOnly(doc.getHtml());
+			return;
+		} catch (AplicacaoException | SwaggerException e) {
+			throw e;
 		} catch (Exception e) {
 			e.printStackTrace(System.out);
 			throw e;
 		}
-		
-		/* TODO: Acesso atraves do login no siga */
-		
-//		SwaggerHelper.buscarEValidarUsuarioLogado();
-//		SigaObjects so = SwaggerHelper.getSigaObjects();
-//		so.assertAcesso("DOC:Módulo de Documentos;" + "");
-//
-//		try {
-//			DpPessoa cadastrante = so.getCadastrante();
-//			DpPessoa titular = cadastrante;
-//			DpLotacao lotaTitular = cadastrante.getLotacao();
-//
-//			ExMobilDaoFiltro flt = new ExMobilDaoFiltro();
-//			flt.setSigla(req.sigla);
-//			ExMobil mob = ExDao.getInstance().consultarPorSigla(flt);
-//			
-//			Utils.assertAcesso(mob, titular, lotaTitular);
-//			ExDocumento doc = mob.doc();
-//			resp.html = doc.getHtml();
-//		} catch (Exception e) {
-//			e.printStackTrace(System.out);
-//			throw e;
-//		}
 	}
 
 	@Override
