+++ sigaex/src/main/webapp/WEB-INF/web.xml
@@ -111,14 +111,6 @@
 		<filter-name>ExFilter</filter-name>
 		<url-pattern>/webdav/*</url-pattern>
 		<url-pattern>/apis/*</url-pattern>
-		
-		<!-- Filter API -->
-		<url-pattern>/api/v1/test/*</url-pattern>
-		<url-pattern>/api/v1/documentos/*</url-pattern>
-		<url-pattern>/api/v1/mesa/*</url-pattern>
-		<url-pattern>/api/v1/acessos/*</url-pattern>
-		<url-pattern>/api/v1/token/*</url-pattern>
-
 	</filter-mapping>
 
 	<filter-mapping>
